consumer/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
backoff==2.2.1
click==8.3.0
confluent-kafka==2.11.1
fastapi==0.117.1
grpcio==1.75.1
grpcio-health-checking==1.75.1
grpcio-tools==1.75.0
h11==0.16.0
httptools==0.6.4
idna==3.10
kafka-python==2.2.15
packaging==25.0
protobuf==6.32.1
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
